LayoutTests/imported/w3c/web-platform-tests/css/css-values/tree-counting/sibling-index-keyframe-rotate-dynamic-expected.txt


PASS Initially, the sibling-index() is 3 for #target
FAIL Removing a preceding sibling of #target reduces the sibling-index() assert_equals: expected "x 20deg" but got "x 30deg"
